refactor: domain 단위로 관리를 위해 domain 상위 패키지 생성 및 기존 user 패키지 이동
