+++ src/main/java/com/lgmpjt/websnsspringboot/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.lgmpjt.websnsspringboot.user.repository;
+package com.lgmpjt.websnsspringboot.domain.user.repository;
 
-import com.lgmpjt.websnsspringboot.user.model.User;
+import com.lgmpjt.websnsspringboot.domain.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
