+++ src/main/java/com/lgmpjt/websnsspringboot/domain/user/service/UserService.java
@@ -1,10 +1,10 @@
-package com.lgmpjt.websnsspringboot.user.service;
+package com.lgmpjt.websnsspringboot.domain.user.service;
 
+import com.lgmpjt.websnsspringboot.domain.user.data.UserSearchUpdateDto;
+import com.lgmpjt.websnsspringboot.domain.user.model.User;
 import com.lgmpjt.websnsspringboot.encryption.service.SHA256;
 import com.lgmpjt.websnsspringboot.mapper.user.UserMapper;
-import com.lgmpjt.websnsspringboot.user.data.UserCreateDto;
-import com.lgmpjt.websnsspringboot.user.data.UserSearchUpdateDto;
-import com.lgmpjt.websnsspringboot.user.model.User;
+import com.lgmpjt.websnsspringboot.domain.user.data.UserCreateDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
